Ajustar permissões

+++ src/main/java/com/lmlasmo/forum/controller/register/ResponseRegisterController.java
@@ -1,6 +1,7 @@
 package com.lmlasmo.forum.controller.register;
 
 import org.springframework.http.ResponseEntity;
+import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
@@ -11,6 +12,8 @@ import org.springframework.web.bind.annotation.RestController;
 
 import com.lmlasmo.forum.dto.generic.ResponseDTO;
 import com.lmlasmo.forum.dto.register.RResponseDTO;
+import com.lmlasmo.forum.model.RoleType;
+import com.lmlasmo.forum.model.Users;
 import com.lmlasmo.forum.service.ResponseService;
 
 import jakarta.validation.Valid;
@@ -33,28 +36,45 @@ public class ResponseRegisterController {
 		return ResponseEntity.ok(response);
 	}
 	
-	@DeleteMapping("/delete/{id}")
+	@DeleteMapping("/delete/{id}")	
 	public ResponseEntity<Object> delete(@PathVariable("id") int id){
 		
-		boolean deleted = service.deleteById(id);
+		ResponseDTO response = service.findById(id);
 		
-		if(deleted) {
-			return ResponseEntity.ok().build();
+		Users user = (Users) SecurityContextHolder.getContext().getAuthentication().getPrincipal(); 
+		
+		if(response.getAuthor() == user.getId() || user.getRole().getRole().equals(RoleType.USER_ADMIN)) {
+			
+			boolean deleted = service.deleteById(id);
+			
+			if(deleted) {
+				return ResponseEntity.ok().build();
+			}
+			
+			return ResponseEntity.badRequest().build();
 		}
 		
-		return ResponseEntity.badRequest().build();
+		return ResponseEntity.status(403).build();				
 	}
 	
-	@PutMapping("/solved/set/{id}")
+	@PutMapping("/solved/set/{id}")	
 	public ResponseEntity<ResponseDTO> setSolved(@PathVariable("id") int id){
 		
-		ResponseDTO response = service.setSolved(id);
+		ResponseDTO response = service.findById(id);
+		Users user = (Users) SecurityContextHolder.getContext().getAuthentication().getPrincipal();
 		
-		if(response != null) {
-			return ResponseEntity.ok(response);
+		if(response.getAuthor() == user.getId()) {
+			
+			ResponseDTO dto = service.setSolved(id);
+			
+			if(response != null) {
+				return ResponseEntity.ok(dto);
+			}
+			
+			return ResponseEntity.notFound().build();									
 		}
 		
-		return ResponseEntity.notFound().build();
+		return ResponseEntity.status(403).build();				
 	}
 	
 }
