+++ src/main/java/com/lmlasmo/forum/controller/register/TopicRegisterController.java
@@ -2,6 +2,7 @@ package com.lmlasmo.forum.controller.register;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.ResponseEntity;
+import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
@@ -11,6 +12,8 @@ import org.springframework.web.bind.annotation.RestController;
 
 import com.lmlasmo.forum.dto.generic.TopicDTO;
 import com.lmlasmo.forum.dto.register.RTopicDTO;
+import com.lmlasmo.forum.model.RoleType;
+import com.lmlasmo.forum.model.Users;
 import com.lmlasmo.forum.service.TopicService;
 
 import jakarta.validation.Valid;
@@ -26,8 +29,11 @@ public class TopicRegisterController {
 		this.service = service;
 	}
 	
-	@PostMapping("/register")
+	@PostMapping("/register")	
 	public ResponseEntity<TopicDTO> register(@RequestBody @Valid RTopicDTO dto){
+		
+		Users user = (Users) SecurityContextHolder.getContext().getAuthentication().getPrincipal(); 		
+		dto.setAuthor(user.getId());
 		
 		TopicDTO topic = service.save(dto);
 		
@@ -37,13 +43,21 @@ public class TopicRegisterController {
 	@DeleteMapping("/delete/{id}")
 	public ResponseEntity<Object> delete(@PathVariable("id") int id){
 		
-		boolean deleted = service.deleteById(id);
+		Users user = (Users) SecurityContextHolder.getContext().getAuthentication().getPrincipal();	
+		TopicDTO topic = service.findById(id);
 		
-		if(deleted) {
-			return ResponseEntity.ok().build();
-		}
+		if(topic.getAuthor() == topic.getId() || user.getRole().getRole().equals(RoleType.USER_ADMIN)) {
+			
+			boolean deleted = service.deleteById(id);
+			
+			if(deleted) {
+				return ResponseEntity.ok().build();
+			}
+			
+			return ResponseEntity.badRequest().build();			
+		}		
 		
-		return ResponseEntity.badRequest().build();
+		return ResponseEntity.status(403).build();	
 	}
 	
 }
